更新20210803

+++ pythonscript/AutoCode/AutoCode.py
@@ -1,11 +1,28 @@
 import os
 from posixpath import dirname
+from re import T
 import sys
 import json
 from subprocess import Popen
 from typing import List
+# sys.path.append("..")
+# from commonfun import *
 
 configFilePath = os.path.abspath(__file__)
+
+def commadEqual(f,t):
+    if f == "-"+t or f == "--"+t or f == t:
+        return True
+    return False
+
+def commadsEqual(inputCmd,cmds):
+    if str(type(cmds)).find("list") !=-1:
+        for cmd in cmds:
+            if(commadEqual(cmd,inputCmd)):
+                return True
+        return False
+    else:
+        return commadEqual(inputCmd,cmds)
 
 def getConifgDir(type,config):
     cfg = os.path.dirname(config)
@@ -25,16 +42,144 @@ def getClassName(fileName):
             className += c
     return className
 
-def getNewContent(ori,oriStr,rStr):
-    newContent = ori.replace(f'{oriStr}',rStr)
-    newContent = newContent.replace(f'{oriStr.upper()}',rStr.upper())
-    newContent = newContent.replace(f'{oriStr.lower()}',rStr.lower())
-    return newContent
-
 def getJScontent(dirPath):
     configJson = f'{dirPath}/config.json'
     cr = open(configJson, "r")
     return json.load(cr)
+
+def strRepalce(ori,oriStr,rStr):
+    newContent=""
+    index=0
+    try:
+        while index < len(ori):
+            beforeIndex=index
+            index = ori.index(oriStr,index)
+            newContent+=ori[beforeIndex:index]
+            nextStr=ori[index:]
+            jStr=nextStr[len(oriStr):].strip()
+            subtraction="{-s}"
+            subtractionEnd="{-e}"
+            if jStr.startswith(subtraction):
+                temp=jStr.replace(subtraction,"").strip()
+                temp=temp.split(subtractionEnd)[0]
+                index=ori.index(temp,index)+len(temp)+len(subtractionEnd)
+                temp=rStr.replace(temp,"")
+                newContent+=temp
+            else:
+                index+=len(oriStr)
+                newContent+=rStr
+    except:
+        if len(newContent) ==0:
+            return ori
+        else:
+            newContent+=ori[index:]
+            return newContent
+    return newContent
+
+def getNewContent(ori,oriStr,rStr):
+    # newContent = ori.replace(oriStr,rStr)
+    # newContent = newContent.replace(oriStr.upper(),rStr.upper())
+    # newContent = newContent.replace(oriStr.lower(),rStr.lower())
+    newContent =strRepalce(ori,oriStr,rStr)
+    newContent =strRepalce(newContent,oriStr.upper(),rStr.upper())
+    newContent =strRepalce(newContent,oriStr.lower(),rStr.lower())
+    return newContent
+
+def contentRepalce(content,replaceList,className,fileName):
+    newContent = getNewContent(content,"{ClassName}",className)
+    newContent = getNewContent(newContent,"{fileName}",fileName)
+    #参数名和对应参数替换
+    index=1
+    for replacestr in replaceList:
+        newContent = getNewContent(newContent,f'{{{index}}}',replacestr)
+        index+=1
+    return newContent
+
+def analylineBrackets(content):
+    num=0
+    for char in content:
+        if char == '{':
+            num+=1
+        elif char == "}":
+            num-=1
+    return num
+
+def modifyFile(jsConfig,replaceList,className,fileName):
+
+    try:
+        modifyDist=jsConfig["M"]
+    except:
+        return
+
+    for contentValue in modifyDist:
+        contents=[]
+        text=contentValue["content"]
+        if type(text) == str:
+            contents.append(text)
+        else:
+            contents=text
+        newContentlist=[]
+        for index,content in enumerate(contents):
+            newContentlist += contentRepalce(content,replaceList,className,fileName)
+            if index != len(contents)-1:
+                newContentlist+="\n"
+
+        path=contentValue["path"]
+        pos=contentValue["pos"]
+
+        #pos 为空，就在最后添加
+        # if len(pos) == 0:
+        #     w_in = open(path,"a")
+        #     w_in.write("\n")
+        #     w_in.writelines(newContentlist)
+        #     w_in.close()
+        #     continue
+
+        r_in=open(path)
+        lineContents=r_in.readlines()
+        r_in.close()
+        w_in = open(path,"w")
+        isStart=False
+        isFinish=False
+        bracketsNum=-1
+        #找到修改的位置
+        for lineContent in lineContents:
+            #如果没有写入就找位置，已经写入就不找位置
+            if not isFinish:
+                #找开始的位置
+                if not isStart:
+                    if pos in lineContent and ";" not in lineContent:
+                        isStart = True
+
+                if isStart:
+                    #找开始括号的位置
+                    #出现括号的位置开始统计左括号的位置
+                    if bracketsNum == -1:
+                        if "{" in lineContent:
+                            bracketsNum=0
+                            bracketsNum+=analylineBrackets(lineContent)
+                    else:
+                        bracketsNum+=analylineBrackets(lineContent)
+                    
+                    #写入内容
+                    if bracketsNum==0:
+                        index=lineContent.find("}")
+                        w_in.write(lineContent[0:index])
+                        w_in.writelines(newContentlist)
+                        w_in.write("\n")
+                        w_in.write(lineContent[index:])
+                        isFinish=True
+                        continue
+
+            w_in.write(lineContent)
+
+        #如果没有找到，就在最后添加
+        if not isFinish:
+            w_in.write("\n")
+            w_in.writelines(newContentlist)
+        print(f"修改 {path} 完成")
+        w_in.close()
+                
 
 def createfile(fileName,type,replaceList):
     configDir = getConifgDir(type,configFilePath)
@@ -42,8 +187,9 @@ def createfile(fileName,type,replaceList):
         print("没有此类型！")
         return
     className = getClassName(fileName)
+    print(f'class {className}')
     for (dirpath,dirnames,filenames) in os.walk(configDir):
-        aimPaths = getJScontent(dirpath)
+        jsConfig = getJScontent(dirpath)
 
         for oriName in filenames:
             if oriName=="config.json":
@@ -52,8 +198,10 @@ def createfile(fileName,type,replaceList):
 
             #得到目标文件路径
             suffix = os.path.splitext(oriName)[1]
-            if suffix in aimPaths:
-                aimPath = aimPaths[suffix]
+            if(suffix.find("h") !=-1 ):
+                print(f'#include "{fileName}{suffix}"')
+            if suffix in jsConfig:
+                aimPath = jsConfig[suffix]
             else:
                 aimPath = os.getcwd()
 
@@ -61,8 +209,14 @@ def createfile(fileName,type,replaceList):
             if  oriName.find("{fileName}") != -1:
                 aimPath = f'{aimPath}/{oriName.replace("{fileName}",fileName)}'
             else:
-                aimPath = f'{aimPath}/{fileName}{suffix}'
+               aimPath = f'{aimPath}/{fileName}{suffix}'
 
+            if os.path.exists(aimPath) :
+                print("文件已经存在，是否替换(y/n)？")
+                isReplce = input()
+                if isReplce == 'n':
+                    continue
+                
             r = open(oriPath,"r")
             w = open(aimPath,"w")
             contents = r.readlines()
@@ -70,57 +224,75 @@ def createfile(fileName,type,replaceList):
 
             #进行替换
             for content in contents:
-                newContent = getNewContent(content,"{ClassName}",className)
-                newContent = getNewContent(newContent,"{fileName}",fileName)
-                #参数名和对应参数替换
-                index=0
-                for replacestr in replaceList:
-                    newContent = getNewContent(newContent,f'{{{index}}}',replacestr)
-                    index+=1
-
+                newContent = contentRepalce(content,replaceList,className,fileName)
                 w.write(newContent)
-
             w.flush()
             w.close()
             print("生成  "+aimPath+'    文件')
+
+    #通过 config 修改文件
+    modifyFile(jsConfig,replaceList,className,fileName)
+
     print("生存完成!")
 
 def showType():
+    if len(sys.argv) == 2 :
+        commad="t"
+        if commadsEqual(sys.argv[1],commad):
+            dirPath = os.path.dirname(configFilePath)
+            index=1
+            for dirname in os.listdir(dirPath):
+                typePath = dirPath+'/'+dirname
+                if(os.path.isdir(typePath)):
+                    js = getJScontent(typePath)
+                    if "desc" in js:
+                        print(f'{str(index):<6}{dirname:<35}'+js["desc"])
+                    else:
+                        print(f'{str(index):<6}{dirname:<35}')
+                index+=1
+            exit()
+
+def getTypeByIndex(currnetType):
     dirPath = os.path.dirname(configFilePath)
+    index=1
     for dirname in os.listdir(dirPath):
         typePath = dirPath+'/'+dirname
         if(os.path.isdir(typePath)):
-            js = getJScontent(typePath)
-            if "desc" in js:
-                print(f'{dirname:<25}'+js["desc"])
-            else:
-                print(dirname)
+            if(index == currnetType):
+                return dirname
+        index+=1
+    
 
-def userVSCodeOpenType(type):
-    configDir = getConifgDir(type,configFilePath)
-    if not(os.path.isdir(configDir)):
-        print("没有此类型！")
-        return
-    proc = Popen(
-        args=f'code {configDir}',
-        shell=True
-    )
+def userVSCodeOpenType():
+    commad=["c","code"]
+    if len(sys.argv) >= 2 :
+        if commadsEqual(sys.argv[1],commad):
+            try:
+                type=sys.argv[2]
+            except:
+                type=""
+            configDir = getConifgDir(type,configFilePath)
+            if not(os.path.isdir(configDir)):
+                print("没有此类型！")
+                return
+            proc = Popen(
+                args=f'code {configDir}',
+                shell=True
+            )
+            exit()
 
-
+# createfile("vc_test","vehctrl",["CDC_2D2__CdcTotMilg","TOPIC_VEHCTRL_CDC_TOT_MILG","测试","float"])
 if __name__ == "__main__":
-    if len(sys.argv) == 2 :
-        if sys.argv[1] == "-s" or sys.argv[1] == 's':
-            showType()
-            exit()
-        elif sys.argv[1] == "-c" or sys.argv[1] == 'code':
-            userVSCodeOpenType("")
-            exit()
+    if len(sys.argv) == 1 :
+        print("请输入:AutoCode 类型 类名，如")
+        print("AutoCode vehctrl vc_discharge_soc CDC_2D2__CdcDchaSocLimit TOPIC_VEHCTRL_CDC_DCHA_SOC_LIMIT 交流放电截止SOC设置请求")
+        exit()
 
-    if len(sys.argv) == 3:
-        if sys.argv[1] == "code" or sys.argv[1] == "c":
-            userVSCodeOpenType(sys.argv[2])
-            exit()
-
+    showType()
+    userVSCodeOpenType()
+    if commadEqual(sys.argv[1],"help"):
+        print(open( os.path.dirname(configFilePath)+"/README.txt").read())
+    
     if len(sys.argv) < 2:
         print("请输入文件名称")
         exit()
@@ -132,4 +304,12 @@ if __name__ == "__main__":
     del replaceList[0]
     del replaceList[0]
     del replaceList[0]
-    createfile(sys.argv[2], sys.argv[1],replaceList)
+    typename=sys.argv[1]
+
+    try:
+        if(type(int(sys.argv[1])) == int):
+            typename=getTypeByIndex(int(sys.argv[1]))
+    except:
+        typename=sys.argv[1]
+    print(replaceList)
+    createfile(sys.argv[2], typename,replaceList)
